Extract shared WindowsFirmware logic into common base class
Move WMI query logic and memoized getters into a common abstract
WindowsFirmware class in oshi-common. JNA and FFM subclasses now
only provide the WmiQueryExecutor.

Partial fix for #3274 (Issue 1: JNA/FFM parallel implementations)
