+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsFirmwareJNA.java
@@ -4,72 +4,21 @@
  */
 package oshi.hardware.platform.windows;
 
-import static oshi.util.Memoizer.memoize;
-
-import java.util.function.Supplier;
+import java.util.Objects;
 
 import oshi.annotation.concurrent.Immutable;
-import oshi.driver.common.windows.wmi.Win32Bios.BiosProperty;
-import oshi.driver.common.windows.wmi.WmiResult;
-import oshi.driver.common.windows.wmi.WmiUtil;
-import oshi.driver.windows.wmi.Win32BiosJNA;
-import oshi.hardware.common.AbstractFirmware;
-import oshi.util.Constants;
-import oshi.util.Util;
-import oshi.util.tuples.Quintet;
+import oshi.driver.common.windows.wmi.WmiQueryExecutor;
+import oshi.hardware.common.platform.windows.WindowsFirmware;
+import oshi.util.platform.windows.WmiQueryExecutorJNA;
 
 /**
- * Firmware data obtained from WMI
+ * Firmware data obtained from WMI using JNA.
  */
 @Immutable
-final class WindowsFirmwareJNA extends AbstractFirmware {
-
-    private final Supplier<Quintet<String, String, String, String, String>> manufNameDescVersRelease = memoize(
-            WindowsFirmwareJNA::queryManufNameDescVersRelease);
+final class WindowsFirmwareJNA extends WindowsFirmware {
 
     @Override
-    public String getManufacturer() {
-        return manufNameDescVersRelease.get().getA();
-    }
-
-    @Override
-    public String getName() {
-        return manufNameDescVersRelease.get().getB();
-    }
-
-    @Override
-    public String getDescription() {
-        return manufNameDescVersRelease.get().getC();
-    }
-
-    @Override
-    public String getVersion() {
-        return manufNameDescVersRelease.get().getD();
-    }
-
-    @Override
-    public String getReleaseDate() {
-        return manufNameDescVersRelease.get().getE();
-    }
-
-    private static Quintet<String, String, String, String, String> queryManufNameDescVersRelease() {
-        String manufacturer = null;
-        String name = null;
-        String description = null;
-        String version = null;
-        String releaseDate = null;
-        WmiResult<BiosProperty> win32BIOS = Win32BiosJNA.queryBiosInfo();
-        if (win32BIOS.getResultCount() > 0) {
-            manufacturer = WmiUtil.getString(win32BIOS, BiosProperty.MANUFACTURER, 0);
-            name = WmiUtil.getString(win32BIOS, BiosProperty.NAME, 0);
-            description = WmiUtil.getString(win32BIOS, BiosProperty.DESCRIPTION, 0);
-            version = WmiUtil.getString(win32BIOS, BiosProperty.VERSION, 0);
-            releaseDate = WmiUtil.getDateString(win32BIOS, BiosProperty.RELEASEDATE, 0);
-        }
-        return new Quintet<>(Util.isBlank(manufacturer) ? Constants.UNKNOWN : manufacturer,
-                Util.isBlank(name) ? Constants.UNKNOWN : name,
-                Util.isBlank(description) ? Constants.UNKNOWN : description,
-                Util.isBlank(version) ? Constants.UNKNOWN : version,
-                Util.isBlank(releaseDate) ? Constants.UNKNOWN : releaseDate);
+    protected WmiQueryExecutor getWmiQueryExecutor() {
+        return Objects.requireNonNull(WmiQueryExecutorJNA.createInstance());
     }
 }
